+++ database/migrations/2026_07_27_110446_create_document_types_table.php
@@ -11,12 +11,14 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('document_types', function (Blueprint $table) {
+        if (!Schema::hasTable('document_types')) {
+            Schema::create('document_types', function (Blueprint $table) {
             $table->id();
             $table->foreignId('company_id');
             $table->string('document_type');
             $table->foreign('company_id')->references('id')->on('companies')->onDelete('cascade');
         });
+        }
     }
 
     /**
